Allow Get Route Contract to have empty response.
We want to use a 204 on a Get endpoint here:

https://github.com/lokalise/autopilot/pull/5664/files#diff-b568a28d846a8c953d129e18f5ed874d5f9a442d970c4b85f419f06e879f0278R498

We believe this is a semantically correct use of this status code, it will be provided when there is no content and we don't expect there to be any content. This is distinct from `data: []` which could be empty for another reason, like the import hasn't begun, or there was some unknown issue.

However our API contracts don't allow get routes to be empty, and give a type error trying to build a fastify route like this, so proposing loosening this type. I'm open to this being challenged, but I don't see why we should restrict GET requests to never being empty, the HTTP spec allows it and it makes sense for certain use-cases.